Contractor entity change

+++ Cursa/Controllers/ContractorsController.cs
@@ -1,25 +1,73 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
+using System.Linq.Dynamic.Core;
 using System.Threading.Tasks;
+using AutoMapper;
+using Cursa.ViewModels.ContractorsVM;
+using Cursa.ViewModels.OwnerVM;
 using DataLayer;
 using DataLayer.Entities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace Cursa.Controllers
 {
     public class ContractorsController : Controller
     {
         private readonly EfDbContext _context;
+        private readonly IMapper _mapper;
+        private readonly ILogger<ContractorsController> _logger;
 
-        public ContractorsController(EfDbContext context)
+        public ContractorsController(EfDbContext context, IMapper mapper, ILogger<ContractorsController> logger)
         {
             _context = context;
+            _mapper = mapper;
+            _logger = logger;
         }
 
         // GET: Contractors
-        public async Task<IActionResult> Index()
+        public IActionResult Index() => View(new ContractorsDisplayViewModel());
+
+        [HttpPost]
+        public IActionResult GetOwners()
         {
-            return View(await _context.Contractors.AsNoTracking().ToListAsync());
+            try
+            {
+                var draw = Request.Form["draw"].FirstOrDefault();
+                var start = Request.Form["start"].FirstOrDefault();
+                var length = Request.Form["length"].FirstOrDefault();
+                var sortColumn = Request
+                    .Form["columns[" + Request.Form["order[0][column]"].FirstOrDefault() + "][name]"].FirstOrDefault();
+                var sortColumnDirection = Request.Form["order[0][dir]"].FirstOrDefault();
+                var searchGlobalValue = Request.Form["search[value]"].FirstOrDefault();
+                var pageSize = length != null ? Convert.ToInt32(length) : 0;
+                var skip = start != null ? Convert.ToInt32(start) : 0;
+
+                var projectsData = _mapper.ProjectTo<ContractorsDisplayViewModel>(_context.Contractors
+                    .AsNoTracking());
+
+                if (!(string.IsNullOrEmpty(sortColumn) && string.IsNullOrEmpty(sortColumnDirection)))
+                {
+                    projectsData = projectsData.OrderBy(sortColumn + " " + sortColumnDirection);
+                }
+
+                if (!string.IsNullOrEmpty(searchGlobalValue))
+                {
+                    projectsData = projectsData.Where(m => m.Name.Contains(searchGlobalValue));
+                }
+
+                var recordsTotal = projectsData.Count();
+                var data = projectsData.Skip(skip).Take(pageSize).ToList();
+                var jsonData = new
+                    {draw = draw, recordsFiltered = recordsTotal, recordsTotal = recordsTotal, data = data};
+                return Ok(jsonData);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError("Error search:{ExceptionMessage}", e.Message);
+                return NotFound();
+            }
         }
 
         // GET: Contractors/Details/5
@@ -55,9 +103,28 @@ namespace Cursa.Controllers
         {
             if (ModelState.IsValid)
             {
-                _context.Add(contractor);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (_context.Statuses.Any(x => String.Equals(x.Name, contractor.Name)))
+                {
+                    ModelState.AddModelError("Name", "Уже существует");
+                }
+
+                if (ModelState.IsValid)
+                {
+                    try
+                    {
+                        _context.Add(contractor);
+                        await _context.SaveChangesAsync();
+                        return RedirectToAction(nameof(Index));
+                    }
+                    catch (DbUpdateException e)
+                    {
+                        var exception = e.InnerException;
+                        if (exception != null && exception.Message.Contains("IX_Contractors_Name"))
+                        {
+                            ModelState.AddModelError("Name", "Контрагент уже существует");
+                        }
+                    }
+                }
             }
             return View(contractor);
         }
@@ -96,6 +163,7 @@ namespace Cursa.Controllers
                 {
                     _context.Update(contractor);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -108,7 +176,15 @@ namespace Cursa.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction(nameof(Index));
+                catch (DbUpdateException e)
+                {
+                    var exception = e.InnerException;
+                    if (exception != null && exception.Message.Contains("IX_Contractors_Name"))
+                    {
+                        ModelState.AddModelError("Name", "Контрагент уже существует");
+                    }
+                }
+                
             }
             return View(contractor);
         }
@@ -137,9 +213,22 @@ namespace Cursa.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var contractor = await _context.Contractors.FindAsync(id);
-            _context.Contractors.Remove(contractor);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (contractor == null)
+            {
+                return NotFound();
+            }
+            try
+            {
+                _context.Contractors.Remove(contractor);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }catch (DbUpdateException e)
+            {
+                _logger.LogInformation("{ExceptionMessage}", e.Message);
+                ModelState.AddModelError(String.Empty, "Невозможно удалить, эта запись используется");
+            }
+
+            return View(contractor);
         }
 
         private bool ContractorExists(int id)
